+++ src/components/meme/InfiniteMemeList/index.ts
@@ -1,1 +1,1 @@
-export * from "./InfiniteMemeList";
+export * from "./MasonryInfiniteGrid";
